feat(my_captains): setup customer page to manages the captains
